Extract `ActionOptions` component
