+++ frontend/src/metabase/dashboard/components/ClickBehaviorSidebar/ClickBehaviorSidebar.jsx
@@ -19,7 +19,6 @@ import {
 } from "metabase/lib/click-behavior";
 import { keyForColumn } from "metabase/lib/dataset";
 
-import Actions from "metabase/entities/actions";
 import Dashboards from "metabase/entities/dashboards";
 import Questions from "metabase/entities/questions";
 
@@ -32,6 +31,7 @@ import ClickMappings, {
 } from "metabase/dashboard/components/ClickMappings";
 
 import { clickBehaviorOptions, getClickBehaviorOptionName } from "./utils";
+import ActionOptions from "./ActionOptions";
 import Column from "./Column";
 import CustomLinkText from "./CustomLinkText";
 import LinkOption from "./LinkOption";
@@ -347,82 +347,6 @@ class ClickBehaviorSidebar extends React.Component {
       </Sidebar>
     );
   }
-}
-
-const ActionOption = ({ name, description, isSelected, onClick }) => {
-  return (
-    <SidebarItemWrapper
-      onClick={onClick}
-      style={{
-        ...SidebarItemStyle,
-        backgroundColor: isSelected ? color("brand") : "transparent",
-        color: isSelected ? color("white") : "inherit",
-        alignItems: description ? "flex-start" : "center",
-        marginTop: "2px",
-      }}
-    >
-      <SidebarIconWrapper>
-        <Icon
-          name="bolt"
-          color={isSelected ? color("text-white") : color("brand")}
-        />
-      </SidebarIconWrapper>
-      <div>
-        <h4>{name}</h4>
-        {description && (
-          <span
-            className={isSelected ? "text-white" : "text-medium"}
-            style={{ width: "95%", marginTop: "2px" }}
-          >
-            {description}
-          </span>
-        )}
-      </div>
-    </SidebarItemWrapper>
-  );
-};
-
-function ActionOptions({ dashcard, clickBehavior, updateSettings }) {
-  return (
-    <SidebarContent>
-      <Heading className="text-medium">{t`Pick an action`}</Heading>
-      <Actions.ListLoader>
-        {({ actions }) => {
-          const selectedAction = actions.find(
-            action => action.id === clickBehavior.action,
-          );
-          return (
-            <>
-              {actions.map(action => (
-                <ActionOption
-                  key={action.id}
-                  name={action.name}
-                  description={action.description}
-                  isSelected={clickBehavior.action === action.id}
-                  onClick={() =>
-                    updateSettings({
-                      type: clickBehavior.type,
-                      action: action.id,
-                      emitter_id: clickBehavior.emitter_id,
-                    })
-                  }
-                />
-              ))}
-              {selectedAction && (
-                <ClickMappings
-                  isAction
-                  object={selectedAction}
-                  dashcard={dashcard}
-                  clickBehavior={clickBehavior}
-                  updateSettings={updateSettings}
-                />
-              )}
-            </>
-          );
-        }}
-      </Actions.ListLoader>
-    </SidebarContent>
-  );
 }
 
 function CrossfilterOptions({
